cython_freeze: allow embedding of submodules
CPython does not allow embedding submodules (see issue 1644818).  This
presents a problem if we wish to create an embedded executable that
contains Cython modules as part of a package.  This patch implements
a work-around.  It appears to work in my limited testing, but I am not
declaring that this is the "proper" way to do it.

As an example, say we want to create an embedded executable containing
the module "testlib.some_module".  In the non-embedded (i.e. dynamic
loading) case, this is no problem - CPython looks for
testlib/some_module.so.  However, there is no obvious way tell CPython
that "testlib.some_module" is a built-in module.

A poor workaround is to remove the module from the package and instead
make it a top-level module.  In some cases this may work, but often this
is not feasible.  To do this, change all of the code to import
"some_module" instead of "testlib.some_module", and change the module
name to "some_module" within the C code.

If you try to embed a module with PyImport_ExtendInittab whose name
contains a period (that is, a submodule, say "testlib.some_module"),
CPython will *not* interpret the period specially, and will create a new
built-in module with a period in its name ("testlib.some_module").  This
new module will not be considered part of the package ("testlib"), and
it will not be possible to import this module normally.  For example,
``import testlib.some_module`` will look for "some_module" within the
"testlib" package and will not find it.

However, it *is* possible to access our module, even though it has
a period in its name.  By calling

    mod = imp.load_module("testlib.some_module", None, '', ('', '', 6))

we can initialize our module and get a reference to it.  Then, if we
stick it in ``sys.modules``, we can import it like normal.  There is
just one small condition - we have to import the package beforehand
(``import testlib`` before the above), or else a dummy package will be
created, hampering any attempt to import other submodules.

One downside to the above is that the module will be initialized right
away.  For most cases, this is fine, but a more general solution is to
use a PEP 302 import hook.  In this way, we can intercept any request to
load "testlib.some_module" and perform the above.  The advantage here is
that (a) we don't have to explicitly import the package first, and (b)
the module will only be initialized when needed.  The downside is that
our hook gets called on every import.

This commit implements the PEP 302 approach.  In the future, we may want
to implement the former approach as an option, for cases where the
import hook is undesirable.